Missed one cast.

+++ cpp/non_iid/non_iid_test_run.h
@@ -3,7 +3,7 @@
 
 #include <string>
 #include <vector>
-#include <jsoncpp/json/json.h>
+#include <json/json.h>
 
 #include "../shared/test_run_base.h"
 #include "non_iid_test_case.h"
@@ -17,7 +17,7 @@ public:
         json["IID"] = IID;
 
         Json::Value testCasesJson;
-        for (int i = 0; i < testCases.size(); i++){
+        for (int i = 0; i < (int)testCases.size(); i++){
             testCasesJson[i] = testCases[i].GetAsJson();
         }
